Extract 'ContentDescription' in own file
