+++ Collect/routes/details.ts
@@ -3,13 +3,14 @@
  */
 import express = require('express');
 import download = require('../tools/download');
+import cd = require('../tools/ContentDescription');
 const router = express.Router();
 
 // Get the details page for an item
 router.get('/:id?', (req: express.Request, res: express.Response, next: express.NextFunction) => {
     var id = req.params.id;
 
-    download.ContentDescription.getById(id, function (err, item) {
+    cd.ContentDescription.getById(id, function (err, item) {
         if (err) {
             err['status'] = 500;
             err['api'] = false;
@@ -25,7 +26,7 @@ router.get('/:id?', (req: express.Request, res: express.Response, next: express.
 // We get an id and the name of the button that was pressed
 router.post('/:id?', (req: express.Request, res: express.Response, next: express.NextFunction) => {
     var id = req.params.id;
-    download.ContentDescription.getById(id, function (err, item) {
+    cd.ContentDescription.getById(id, function (err, item) {
         if (err) {
             err['status'] = 500;
             err['api'] = false;
@@ -35,7 +36,7 @@ router.post('/:id?', (req: express.Request, res: express.Response, next: express
         if (req.body.submit !== undefined && req.body.submit != null) {
             // We can't do anything without an title
             if (req.body.title) {
-                download.ContentDescription.setTitle(item.id, req.body.title, function (err, item) {
+                cd.ContentDescription.setTitle(item.id, req.body.title, function (err, item) {
                     if (err) {
                         err['status'] = 500;
                         err['api'] = false;
@@ -55,7 +56,7 @@ router.post('/:id?', (req: express.Request, res: express.Response, next: express
             }
         } // User pressed the 'delete' button
         else if (req.body.delete !== undefined && req.body.delete != null) {
-            download.ContentDescription.removeContent(id, function (err) {
+            cd.ContentDescription.removeContent(id, function (err) {
                 if (err) {
                     err['status'] = 500;
                     err['api'] = false;
