+++ Collect/routes/sites.js
@@ -4,12 +4,12 @@ Object.defineProperty(exports, "__esModule", { value: true });
  * Site page
  */
 const express = require("express");
-const download = require("../tools/download");
+const cd = require("../tools/ContentDescription");
 const router = express.Router();
 //Show all archived pages of one domain
 router.get('/:domain?', (req, res, next) => {
     var domain = req.params.domain;
-    download.ContentDescription.getSaved(function (err, result) {
+    cd.ContentDescription.getSaved(function (err, result) {
         if (err) {
             return res.render('error', { error: err });
         }
